+++ Microarray/Affymetrix/Workflow_Documentation/NF_MAAffymetrix/workflow_code/modules/GENERATE_SOFTWARE_TABLE/resources/usr/bin/SoftwareYamlToMarkdownTable.py
@@ -43,13 +43,23 @@ HOMEPAGE_MAP = {
 
 @click.command()
 @click.argument("input_yaml", type=click.Path(exists=True))
-def yamlToMarkdown(input_yaml: Path):
+@click.argument("filename")
+@click.argument("skip_de", type=click.BOOL)
+def yamlToMarkdown(input_yaml: Path, filename: str, skip_de: bool):
     """ Using a software versions """
     with open(input_yaml, "r") as f:
         data = yaml.safe_load(f)
 
     data.extend(ASSUMED_SOFTWARE)
     df = pd.DataFrame(data)
+
+    # If data files are not compressed, won't use R.utils to unzip them during processing
+    if not filename.endswith('.gz'):
+        AFFYMETRIX_SOFTWARE_DPPD.remove('r.utils')
+
+    if skip_de:
+        AFFYMETRIX_SOFTWARE_DPPD.remove('limma')
+        AFFYMETRIX_SOFTWARE_DPPD.remove('statmod')
 
     # Filter to direct software used (i.e. exclude dependencies of the software)
     df = df.loc[df["name"].str.lower().isin(AFFYMETRIX_SOFTWARE_DPPD)]